Make Nadybot pass strictest psalm settings (#496)
* Skip PPC tests for now
* Don't log notices during CI/CD tests

+++ src/Modules/SKILLS_MODULE/BuffPerksController.php
@@ -485,7 +485,7 @@ class BuffPerksController extends ModuleInstance {
 			name: $perk->name,
 			description: $perk->description,
 			professions: $perk->levels[$minLevel]->professions,
-			max_level: collect($perk->levels)->keys()->max(),
+			max_level: (int)collect($perk->levels)->keys()->max(),
 		);
 
 		/** @var array<int,int> */
